moved index_pipeline and removed dubble classification
